Reference

Contents

Reference

Contents

Index

Index

SPJFractals.axes_bounds

SPJFractals.basic_plot

SPJFractals.burningship_equation

SPJFractals.compute_fractal

SPJFractals.compute_fractal_cpu

SPJFractals.compute_fractal_gpu

SPJFractals.create_newton_equation

SPJFractals.fractal_app

SPJFractals.fractal_closure

SPJFractals.gpu_fractal_kernel!

SPJFractals.julia_equation

SPJFractals.mandelbrot_equation

SPJFractals.plane_ranges

SPJFractals.tricorn_equation

SPJFractals.axes_bounds — Method

axes_bounds(center, plane_size)

Compute the bounds of the axes for a heatmap.

Arguments

center::AbstractVector{<:Real}: The center of the complex plane in image.

plane_size::AbstractVector{<:Real}: The size of the complex plane.

Returns

A tuple of two tuples containing the x and y bounds of the axes.

SPJFractals.basic_plot — Method

compute_fractal(func, params, center, plane_size, img_size, gpu)

Compute a fractal using the given function and parameters.

Arguments

func::Function: The function to compute the fractal with.

params::AbstractArray{<:Number}: The parameters to pass to the function.

center::AbstractVector{<:Real}: The center of the complex plane in image.

plane_size::AbstractVector{<:Real}: The size of the complex plane.

img_size::AbstractVector{<:Integer}: The size of the image.

gpu::Bool: Whether to use the GPU for computation.

Returns

An figure and image containing the computed fractal.

SPJFractals.burningship_equation — Method

burning_ship_equation(x::AbstractFloat, y::AbstractFloat, params::AbstractVector{<:Real})

Compute the number of iterations for the Burning Ship fractal equation.

Arguments

x::AbstractFloat: The x-coordinate of the point.

y::AbstractFloat: The y-coordinate of the point.

params::AbstractVector{<:Real}: Number of iterations.

Returns

The number of iterations before the point escapes.

SPJFractals.compute_fractal — Function

compute_fractal(func, params, center, plane_size, img_size, use_gpu=true)

Compute a fractal using the given function and parameters.

Arguments

func::Function: The function to compute the fractal with.

params::AbstractArray{<:Number}: The parameters to pass to the function.

center::AbstractVector{<:Real}: The center of the complex plane in image.

plane_size::AbstractVector{<:Real}: The size of the complex plane.

img_size::AbstractVector{<:Integer}: The size of the image.

use_gpu::Bool=true: Whether to use the GPU or threaded CPU for computation.

Returns

An Array{UInt16,2} containing the computed fractal.

SPJFractals.compute_fractal_cpu — Method

compute_fractal_cpu(func, params, center, plane_size, img_size)

Compute a fractal using the given function and parameters on the CPU.

Arguments

func::Function: The function to compute the fractal with.

params::AbstractArray{<:Number}: The parameters to pass to the function.

center::AbstractVector{<:Real}: The center of the complex plane in image.

plane_size::AbstractVector{<:Real}: The size of the complex plane.

img_size::AbstractVector{<:Integer}: The size of the image.

Returns

An Array{UInt16,2} containing the computed fractal.

SPJFractals.compute_fractal_gpu — Method

compute_fractal_gpu(func, params, center, plane_size, img_size)

Compute a fractal using the given function and parameters on the GPU.

Arguments

func::Function: The function to compute the fractal with.

params::AbstractArray{<:Number}: The parameters to pass to the function.

center::AbstractVector{<:Real}: The center of the complex plane in image.

plane_size::AbstractVector{<:Real}: The size of the complex plane.

img_size::AbstractVector{<:Integer}: The size of the image.

Returns

An Array{UInt16,2} containing the computed fractal.

SPJFractals.create_newton_equation — Method

create_newton_equation(func::Function)

Create a Newton fractal equation from a differentiable function f(x), solving f(x) = 0. Each root is assigned a unique color based on the angle of the complex number.

Arguments

func::Function: The function to create the Newton fractal equation from.

Returns

A function with input parameters x, y, and params, that computes the Newton fractal.

SPJFractals.fractal_app — Function

fractal_app(img_size::AbstractArray{<:Int} = [1080, 720], fps::Int = 60)

Create an interactive fractal visualization app.

Arguments

img_size::AbstractArray{<:Int}: The size of the image.

fps::Int: The frames per second.

Returns

Nothing.

SPJFractals.fractal_closure — Function

# fractal_closure(func::Function)

Create a closure for computing fractals with the given function and parameters.

Arguments

func::Function: The function to compute the fractal with.

img_size::AbstractVector{<:Integer}: The size of the image.

use_gpu::Bool=true: Whether to use the GPU or threaded CPU for computation.

Returns

A closure that computes a fractal using the given function with parameters: parameters, center, and plane size.

SPJFractals.gpu_fractal_kernel! — Method

gpu_fractal_kernel!(result, func, params, x_vals, y_vals)

A kernel for computing a fractal on the GPU.

Arguments

result::AbstractArray{<:Integer}: The array to store the computed fractal.

func::Function: The function to compute the fractal with.

params::AbstractArray{<:Number}: The parameters to pass to the function.

x_vals::AbstractArray{<:Real}: The x values of the complex plane.

y_vals::AbstractArray{<:Real}: The y values of the complex plane.

SPJFractals.julia_equation — Function

julia_equation(x0::Real, y0::Real, params::AbstractVector{<:Real} = [500, 0.0, 0.0])

Compute the number of iterations for the Julia set equation.

Arguments

x0::Real: The x-coordinate of the point.

y0::Real: The y-coordinate of the point.

params::AbstractVector{<:Real}: Number of iterations, complex part of the constant, and real part of the constant.

Returns

The number of iterations before the point escapes.

SPJFractals.mandelbrot_equation — Method

mandelbrot_equation(x::AbstractFloat, y::AbstractFloat, params::AbstractVector{<:Real})

Compute the number of iterations for the Mandelbrot set equation.

Arguments

x::AbstractFloat: The x-coordinate of the point.

y::AbstractFloat: The y-coordinate of the point.

params::AbstractVector{<:Real}: Number of iterations.

Returns

The number of iterations before the point escapes.

SPJFractals.plane_ranges — Method

plane_ranges(center, size, img_size)

Compute the ranges of the complex plane in image coordinates.

Arguments

center::AbstractVector{<:Real}: The center of the complex plane in image.

size::AbstractVector{<:Real}: The size of the complex plane.

img_size::AbstractVector{<:Integer}: The size of the image.

Returns

A tuple of two arrays containing the x and y ranges of the complex plane.

SPJFractals.tricorn_equation — Method

tricorn_equation(x::AbstractFloat, y::AbstractFloat, params::AbstractVector{<:Real})

Compute the number of iterations for the Tricorn fractal equation.

Arguments

x::AbstractFloat: The x-coordinate of the point.

y::AbstractFloat: The y-coordinate of the point.

params::AbstractVector{<:Real}: Number of iterations.

Returns

The number of iterations before the point escapes.